requirements.txt
# Core Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Classical Forecasting
statsmodels>=0.14.0
prophet>=1.1.0
pmdarima>=2.0.0

# Experiment Tracking
mlflow>=2.5.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pytest>=7.3.0

# Optional: Efficient Attention (if GPU available)
# xformers>=0.0.20
# flash-attn>=2.0.0

